Crate governance_os_pallet_plcr_voting

This pallet can be used to deploy a PLCR voting system. With PLCR voting, votes are hashed before being revealed and counted. This allows people to hide their votes until everyone places theirs as to prevent collusion between participants or typical human biases.

Re-exports

pub use $crate::traits::GetPalletVersion as _;

Structs

Locks
Keeps track of locks set on user’s balances and to which proposal they were linked to.

Module

Proposals
Proposals actively opened and linked to this voting implementation. Erased when closed or vetoed.

Votes
Keeps track of the votes casted in favor of or against the different proposals.

VotingParameters

Enums

Call
Dispatchable calls.

Error

RawEvent
Events for this module.

VoteData

Constants

PLCR_VOTING_LOCK_ID

Traits

Config

Type Definitions

Event
RawEvent specialized for the configuration Config